@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from typing import Optional
 
 import gymnasium as gym
 import numpy as np
@@ -17,7 +16,7 @@ class DummyEnv(gym.Env):
     def step(self, action):
         return self.observation_space.sample(), 0.0, False, False, {}
 
-    def reset(self, *, seed: Optional[int] = None, options: Optional[dict] = None):
+    def reset(self, *, seed: int | None = None, options: dict | None = None):
         if seed is not None:
             super().reset(seed=seed)
         return self.observation_space.sample(), {}
